chore(rust): bump alloy to 2.0
