+++ rust/kona/crates/node/engine/src/test_utils/attributes.rs
@@ -87,6 +87,7 @@ impl TestAttributesBuilder {
                 suggested_fee_recipient: self.suggested_fee_recipient,
                 withdrawals: self.withdrawals,
                 parent_beacon_block_root: self.parent_beacon_block_root,
+                slot_number: Default::default(),
             },
             transactions: self.transactions,
             no_tx_pool: self.no_tx_pool,
