+++ rust/kona/crates/proof/executor/src/test_utils.rs
@@ -151,6 +151,7 @@ impl ExecutorTestFixtureCreator {
                 prev_randao: executing_header.mix_hash,
                 withdrawals: Default::default(),
                 suggested_fee_recipient: executing_header.beneficiary,
+                slot_number: Default::default(),
             },
             gas_limit: Some(executing_header.gas_limit),
             transactions: Some(encoded_executing_transactions),
