+++ rust/kona/crates/protocol/derive/src/attributes/stateful.rs
@@ -211,6 +211,7 @@ where
                 suggested_fee_recipient: Predeploys::SEQUENCER_FEE_VAULT,
                 parent_beacon_block_root: parent_beacon_root,
                 withdrawals,
+                slot_number: None,
             },
             transactions: Some(txs),
             no_tx_pool: Some(true),
@@ -493,6 +494,7 @@ mod tests {
                 suggested_fee_recipient: Predeploys::SEQUENCER_FEE_VAULT,
                 parent_beacon_block_root: None,
                 withdrawals: None,
+                slot_number: None,
             },
             transactions: payload.transactions.clone(),
             no_tx_pool: Some(true),
@@ -545,6 +547,7 @@ mod tests {
                 suggested_fee_recipient: Predeploys::SEQUENCER_FEE_VAULT,
                 parent_beacon_block_root: None,
                 withdrawals: Some(Vec::default()),
+                slot_number: None,
             },
             transactions: payload.transactions.clone(),
             no_tx_pool: Some(true),
@@ -598,6 +601,7 @@ mod tests {
                 suggested_fee_recipient: Predeploys::SEQUENCER_FEE_VAULT,
                 parent_beacon_block_root,
                 withdrawals: Some(vec![]),
+                slot_number: None,
             },
             transactions: payload.transactions.clone(),
             no_tx_pool: Some(true),
@@ -650,6 +654,7 @@ mod tests {
                 suggested_fee_recipient: Predeploys::SEQUENCER_FEE_VAULT,
                 parent_beacon_block_root: Some(B256::ZERO),
                 withdrawals: Some(vec![]),
+                slot_number: None,
             },
             transactions: payload.transactions.clone(),
             no_tx_pool: Some(true),
