+++ rust/kona/crates/protocol/derive/src/pipeline/core.rs
@@ -276,6 +276,7 @@ mod tests {
                     suggested_fee_recipient: Default::default(),
                     withdrawals: None,
                     parent_beacon_block_root: None,
+                    slot_number: None,
                 },
                 transactions: None,
                 no_tx_pool: None,
