+++ rust/kona/crates/protocol/protocol/src/block.rs
@@ -294,6 +294,7 @@ mod tests {
             block_number: Some(1),
             effective_gas_price: Some(1),
             transaction_index: Some(0),
+            block_timestamp: None,
         };
         let block: alloy_rpc_types_eth::Block<op_alloy_rpc_types::Transaction> =
             alloy_rpc_types_eth::Block {
