+++ rust/op-alloy/crates/network/Cargo.toml
@@ -22,7 +22,6 @@ op-alloy-rpc-types.workspace = true
 # Alloy
 alloy-consensus.workspace = true
 alloy-network.workspace = true
-alloy-primitives.workspace = true
 alloy-provider.workspace = true
 alloy-rpc-types-eth.workspace = true
 
@@ -31,13 +30,11 @@ std = [
 	"op-alloy-consensus/std",
 	"op-alloy-rpc-types/std",
 	"alloy-consensus/std",
-	"alloy-primitives/std",
 	"alloy-rpc-types-eth/std"
 ]
 serde = [
 	"op-alloy-consensus/serde",
 	"op-alloy-rpc-types/serde",
 	"alloy-consensus/serde",
-	"alloy-primitives/serde",
 	"alloy-rpc-types-eth/serde"
 ]
