+++ rust/op-alloy/crates/network/src/lib.rs
@@ -9,8 +9,6 @@
 pub use alloy_network::*;
 
 use alloy_consensus::{ReceiptWithBloom, TxType};
-use alloy_primitives::{Address, Bytes, ChainId, TxKind, U256};
-use alloy_rpc_types_eth::AccessList;
 use op_alloy_consensus::{OpReceipt, OpTxType, OpTypedTransaction};
 use op_alloy_rpc_types::OpTransactionRequest;
 
@@ -43,124 +41,28 @@ impl Network for Optimism {
         alloy_rpc_types_eth::Block<Self::TransactionResponse, Self::HeaderResponse>;
 }
 
-impl TransactionBuilder<Optimism> for OpTransactionRequest {
-    fn chain_id(&self) -> Option<ChainId> {
-        self.as_ref().chain_id()
-    }
-
-    fn set_chain_id(&mut self, chain_id: ChainId) {
-        self.as_mut().set_chain_id(chain_id);
-    }
-
-    fn nonce(&self) -> Option<u64> {
-        self.as_ref().nonce()
-    }
-
-    fn set_nonce(&mut self, nonce: u64) {
-        self.as_mut().set_nonce(nonce);
-    }
-
-    fn take_nonce(&mut self) -> Option<u64> {
-        self.as_mut().nonce.take()
-    }
-
-    fn input(&self) -> Option<&Bytes> {
-        self.as_ref().input()
-    }
-
-    fn set_input<T: Into<Bytes>>(&mut self, input: T) {
-        self.as_mut().set_input(input);
-    }
-
-    fn from(&self) -> Option<Address> {
-        self.as_ref().from()
-    }
-
-    fn set_from(&mut self, from: Address) {
-        self.as_mut().set_from(from);
-    }
-
-    fn kind(&self) -> Option<TxKind> {
-        self.as_ref().kind()
-    }
-
-    fn clear_kind(&mut self) {
-        self.as_mut().clear_kind();
-    }
-
-    fn set_kind(&mut self, kind: TxKind) {
-        self.as_mut().set_kind(kind);
-    }
-
-    fn value(&self) -> Option<U256> {
-        self.as_ref().value()
-    }
-
-    fn set_value(&mut self, value: U256) {
-        self.as_mut().set_value(value);
-    }
-
-    fn gas_price(&self) -> Option<u128> {
-        self.as_ref().gas_price()
-    }
-
-    fn set_gas_price(&mut self, gas_price: u128) {
-        self.as_mut().set_gas_price(gas_price);
-    }
-
-    fn max_fee_per_gas(&self) -> Option<u128> {
-        self.as_ref().max_fee_per_gas()
-    }
-
-    fn set_max_fee_per_gas(&mut self, max_fee_per_gas: u128) {
-        self.as_mut().set_max_fee_per_gas(max_fee_per_gas);
-    }
-
-    fn max_priority_fee_per_gas(&self) -> Option<u128> {
-        self.as_ref().max_priority_fee_per_gas()
-    }
-
-    fn set_max_priority_fee_per_gas(&mut self, max_priority_fee_per_gas: u128) {
-        self.as_mut().set_max_priority_fee_per_gas(max_priority_fee_per_gas);
-    }
-
-    fn gas_limit(&self) -> Option<u64> {
-        self.as_ref().gas_limit()
-    }
-
-    fn set_gas_limit(&mut self, gas_limit: u64) {
-        self.as_mut().set_gas_limit(gas_limit);
-    }
-
-    fn access_list(&self) -> Option<&AccessList> {
-        self.as_ref().access_list()
-    }
-
-    fn set_access_list(&mut self, access_list: AccessList) {
-        self.as_mut().set_access_list(access_list);
-    }
-
+impl NetworkTransactionBuilder<Optimism> for OpTransactionRequest {
     fn complete_type(&self, ty: OpTxType) -> Result<(), Vec<&'static str>> {
         match ty {
             OpTxType::Deposit => Err(vec!["not implemented for deposit tx"]),
             _ => {
                 let ty = TxType::try_from(ty as u8).unwrap();
-                self.as_ref().complete_type(ty)
+                NetworkTransactionBuilder::<Ethereum>::complete_type(self.as_ref(), ty)
             }
         }
     }
 
     fn can_submit(&self) -> bool {
-        self.as_ref().can_submit()
+        NetworkTransactionBuilder::<Ethereum>::can_submit(self.as_ref())
     }
 
     fn can_build(&self) -> bool {
-        self.as_ref().can_build()
+        NetworkTransactionBuilder::<Ethereum>::can_build(self.as_ref())
     }
 
     #[doc(alias = "output_transaction_type")]
     fn output_tx_type(&self) -> OpTxType {
-        match self.as_ref().preferred_type() {
+        match NetworkTransactionBuilder::<Ethereum>::output_tx_type(self.as_ref()) {
             TxType::Eip1559 | TxType::Eip4844 => OpTxType::Eip1559,
             TxType::Eip2930 => OpTxType::Eip2930,
             TxType::Eip7702 => OpTxType::Eip7702,
@@ -170,16 +72,18 @@ impl TransactionBuilder<Optimism> for OpTransactionRequest {
 
     #[doc(alias = "output_transaction_type_checked")]
     fn output_tx_type_checked(&self) -> Option<OpTxType> {
-        self.as_ref().buildable_type().map(|tx_ty| match tx_ty {
-            TxType::Eip1559 | TxType::Eip4844 => OpTxType::Eip1559,
-            TxType::Eip2930 => OpTxType::Eip2930,
-            TxType::Eip7702 => OpTxType::Eip7702,
-            TxType::Legacy => OpTxType::Legacy,
+        NetworkTransactionBuilder::<Ethereum>::output_tx_type_checked(self.as_ref()).map(|tx_ty| {
+            match tx_ty {
+                TxType::Eip1559 | TxType::Eip4844 => OpTxType::Eip1559,
+                TxType::Eip2930 => OpTxType::Eip2930,
+                TxType::Eip7702 => OpTxType::Eip7702,
+                TxType::Legacy => OpTxType::Legacy,
+            }
         })
     }
 
     fn prep_for_submission(&mut self) {
-        self.as_mut().prep_for_submission();
+        NetworkTransactionBuilder::<Ethereum>::prep_for_submission(self.as_mut());
     }
 
     fn build_unsigned(self) -> BuildResult<OpTypedTransaction, Optimism> {
