+++ rust/op-alloy/crates/rpc-types-engine/src/attributes.rs
@@ -241,6 +241,7 @@ mod test {
                 suggested_fee_recipient: address!("0x4200000000000000000000000000000000000011"),
                 withdrawals: Some([].into()),
                 parent_beacon_block_root: b256!("0x8fe0193b9bf83cb7e5a08538e494fecc23046aab9a497af3704f4afdae3250ff").into(),
+                slot_number: Default::default(),
             },
             transactions: Some([bytes!("7ef8f8a0dc19cfa777d90980e4875d0a548a881baaa3f83f14d1bc0d3038bc329350e54194deaddeaddeaddeaddeaddeaddeaddeaddead00019442000000000000000000000000000000000000158080830f424080b8a4440a5e20000f424000000000000000000000000300000000670d6d890000000000000125000000000000000000000000000000000000000000000000000000000000000700000000000000000000000000000000000000000000000000000000000000014bf9181db6e381d4384bbf69c48b0ee0eed23c6ca26143c6d2544f9d39997a590000000000000000000000007f83d659683caf2767fd3c720981d51f5bc365bc")].into()),
             no_tx_pool: None,
@@ -274,6 +275,7 @@ mod test {
                 suggested_fee_recipient: address!("0x4200000000000000000000000000000000000011"),
                 withdrawals: Some([].into()),
                 parent_beacon_block_root: b256!("0x8fe0193b9bf83cb7e5a08538e494fecc23046aab9a497af3704f4afdae3250ff").into(),
+                slot_number: Default::default(),
             },
             transactions: Some([bytes!("7ef8f8a0dc19cfa777d90980e4875d0a548a881baaa3f83f14d1bc0d3038bc329350e54194deaddeaddeaddeaddeaddeaddeaddeaddead00019442000000000000000000000000000000000000158080830f424080b8a4440a5e20000f424000000000000000000000000300000000670d6d890000000000000125000000000000000000000000000000000000000000000000000000000000000700000000000000000000000000000000000000000000000000000000000000014bf9181db6e381d4384bbf69c48b0ee0eed23c6ca26143c6d2544f9d39997a590000000000000000000000007f83d659683caf2767fd3c720981d51f5bc365bc")].into()),
             no_tx_pool: None,
@@ -302,6 +304,7 @@ mod test {
                 suggested_fee_recipient: Address::ZERO,
                 withdrawals: Default::default(),
                 parent_beacon_block_root: Some(B256::ZERO),
+                slot_number: Default::default(),
             },
             transactions: Some(vec![b"hello".to_vec().into()]),
             no_tx_pool: Some(true),
@@ -325,6 +328,7 @@ mod test {
                 suggested_fee_recipient: Address::ZERO,
                 withdrawals: Default::default(),
                 parent_beacon_block_root: Some(B256::ZERO),
+                slot_number: Default::default(),
             },
             transactions: Some(vec![b"hello".to_vec().into()]),
             no_tx_pool: Some(true),
@@ -366,6 +370,7 @@ mod test {
                 suggested_fee_recipient: Address::ZERO,
                 withdrawals: Default::default(),
                 parent_beacon_block_root: Some(B256::ZERO),
+                slot_number: Default::default(),
             },
             transactions: Some(vec![b"hello".to_vec().into()]),
             no_tx_pool: Some(true),
@@ -389,6 +394,7 @@ mod test {
                 suggested_fee_recipient: Address::ZERO,
                 withdrawals: Default::default(),
                 parent_beacon_block_root: Some(B256::ZERO),
+                slot_number: Default::default(),
             },
             transactions: Some(vec![b"hello".to_vec().into()]),
             no_tx_pool: Some(true),
