+++ rust/op-alloy/crates/rpc-types/Cargo.toml
@@ -22,6 +22,7 @@ op-alloy-consensus = { workspace = true, features = ["serde"] }
 alloy-serde.workspace = true
 alloy-consensus.workspace = true
 alloy-network-primitives.workspace = true
+alloy-network = { workspace = true, optional = true }
 alloy-eips = { workspace = true, features = ["serde"] }
 alloy-rpc-types-eth = { workspace = true, features = ["serde"] }
 alloy-primitives = { workspace = true, features = ["map", "rlp", "serde"] }
@@ -34,7 +35,6 @@ serde = { workspace = true, features = ["derive"] }
 jsonrpsee = { workspace = true, optional = true }
 reth-rpc-traits = { workspace = true, optional = true }
 alloy-signer = { workspace = true, optional = true }
-alloy-network = { workspace = true, optional = true }
 
 # arbitrary
 arbitrary = { workspace = true, features = ["derive"], optional = true }
@@ -54,6 +54,7 @@ similar-asserts.workspace = true
 [features]
 default = ["std"]
 std = [
+	"dep:alloy-network",
 	"alloy-network-primitives/std",
 	"alloy-eips/std",
 	"alloy-primitives/std",
