+++ rust/op-alloy/crates/rpc-types/src/transaction.rs
@@ -57,6 +57,7 @@ impl<T: OpTransaction + TransactionTrait> Transaction<T> {
                 block_number: tx_info.inner.block_number,
                 transaction_index: tx_info.inner.index,
                 effective_gas_price: Some(effective_gas_price),
+                block_timestamp: tx_info.inner.block_timestamp,
             },
             deposit_nonce: tx_info.deposit_meta.deposit_nonce,
             deposit_receipt_version: tx_info.deposit_meta.deposit_receipt_version,
@@ -251,6 +252,12 @@ mod tx_serde {
             skip_serializing_if = "Option::is_none",
             with = "alloy_serde::quantity::opt"
         )]
+        block_timestamp: Option<u64>,
+        #[serde(
+            default,
+            skip_serializing_if = "Option::is_none",
+            with = "alloy_serde::quantity::opt"
+        )]
         deposit_receipt_version: Option<u64>,
 
         #[serde(flatten)]
@@ -267,6 +274,7 @@ mod tx_serde {
                         block_number,
                         transaction_index,
                         effective_gas_price,
+                        block_timestamp,
                     },
                 deposit_receipt_version,
                 deposit_nonce,
@@ -283,6 +291,7 @@ mod tx_serde {
                 block_hash,
                 block_number,
                 transaction_index,
+                block_timestamp,
                 deposit_receipt_version,
                 other: OptionalFields { from, effective_gas_price, deposit_nonce },
             }
@@ -298,6 +307,7 @@ mod tx_serde {
                 block_hash,
                 block_number,
                 transaction_index,
+                block_timestamp,
                 deposit_receipt_version,
                 other,
             } = value;
@@ -325,6 +335,7 @@ mod tx_serde {
                     block_number,
                     transaction_index,
                     effective_gas_price,
+                    block_timestamp,
                 },
                 deposit_receipt_version,
                 deposit_nonce,
