+++ rust/op-reth/crates/node/src/engine.rs
@@ -337,6 +337,7 @@ mod test {
                 suggested_fee_recipient: Address::ZERO,
                 withdrawals: Some(vec![]),
                 parent_beacon_block_root: Some(B256::ZERO),
+                slot_number: None,
             },
         })
     }
