+++ rust/op-reth/crates/node/tests/e2e-testsuite/testsuite.rs
@@ -40,6 +40,7 @@ async fn test_testsuite_op_assert_mine_block() -> Result<()> {
                     suggested_fee_recipient: Address::random(),
                     withdrawals: None,
                     parent_beacon_block_root: None,
+                    slot_number: None,
                 },
                 transactions: None,
                 no_tx_pool: None,
@@ -85,6 +86,7 @@ async fn test_testsuite_op_assert_mine_block_isthmus_activated() -> Result<()> {
                     suggested_fee_recipient: Address::random(),
                     withdrawals: Some(vec![]),
                     parent_beacon_block_root: Some(B256::ZERO),
+                    slot_number: None,
                 },
                 transactions: None,
                 no_tx_pool: None,
