+++ rust/op-reth/crates/primitives/src/bedrock.rs
@@ -91,6 +91,8 @@ pub const BEDROCK_HEADER: Header = Header {
     excess_blob_gas: None,
     parent_beacon_block_root: None,
     requests_hash: None,
+    block_access_list_hash: None,
+    slot_number: None,
 };
 
 /// Bedrock total difficulty on Optimism Mainnet.
