+++ rust/op-reth/examples/custom-node/src/primitives/header.rs
@@ -135,6 +135,14 @@ impl alloy_consensus::BlockHeader for CustomHeader {
         self.inner.requests_hash()
     }
 
+    fn block_access_list_hash(&self) -> Option<B256> {
+        self.inner.block_access_list_hash()
+    }
+
+    fn slot_number(&self) -> Option<u64> {
+        self.inner.slot_number()
+    }
+
     fn extra_data(&self) -> &Bytes {
         self.inner.extra_data()
     }
